core::Attributes - let's simplify the API and use 'int' instead of 'size_t'
the Attributes container will never have more elements than INT_MAX,
in this way, it is safe to expose the element's access through
simple integers.

+++ src/core/edge_p.cpp
@@ -64,7 +64,7 @@ const Node& Edge::neighbour() const
 const Attributes* Edge::attrs() const
 { return m_ptr->attrs(); }
 
-const Value& Edge::attr(size_t id) const
+const Value& Edge::attr(int id) const
 { return m_ptr->attr(id); }
 
 Value Edge::attr(const QString& name, Value defaultValue) const
